+++ backend/models/users.js
@@ -11,8 +11,8 @@ const addressSchema = mongoose.Schema({
     email: { type: String, default: '' },
     password: { type: String, default: '' },
     token: { type: String, default: '' },
-    name: { type: String, default: '' },
-    surname: { type: String, default: '' },
+    firstname: { type: String, default: '' },
+    lastname: { type: String, default: '' },
     uid: { type: String, default: '' },
     address: { type: addressSchema, default: {} },
     phone: { type: String, default: '' },
