+++ frontend/config.js
@@ -1,5 +1,5 @@
 const config = {
-    URL_BACKEND: "http://192.168.133.233:3000",
+    URL_BACKEND: "http://10.0.2.86:3000",
   };
   
   export default config;
